requirements.txt
torch==1.7
torchvision==0.8.0
opencv==4.4.0.44
numpy==1.21.6
loguru==0.5.3
pandas==1.1.4
Pillow==9.5.0
tensorboardX==2.6.2.2
tqdm>=4.62.2
matplotlib==3.4.3